+++ OUDS/Core/Themes/Orange/Tests/Values/ComponentTokens/ThemeOverrideOfSelectInputComponentTokensTests.swift
@@ -1,0 +1,38 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDS
+import OUDSThemesOrange
+import Testing
+
+// swiftlint:disable type_name
+
+struct ThemeOverrideOfSelectInputComponentTokensTests {
+
+    private var abstractTheme: OUDSTheme
+    private var inheritedTheme: OUDSTheme
+
+    init() {
+        abstractTheme = OrangeTheme()
+        inheritedTheme = MockTheme()
+    }
+
+    // MARK: - Sizes
+
+    @Test func inheritedThemeCanOverrideSelectInputComponentTokenSizeMinWidth() throws {
+        #expect(inheritedTheme.select.selectInputSizeMinWidth != abstractTheme.select.selectInputSizeMinWidth)
+        #expect(inheritedTheme.select.selectInputSizeMinWidth == MockThemeSelectInputComponentTokenProvider.mockThemeSelectInputSize)
+    }
+}
+
+// swiftlint:enable type_name